requirements.txt
streamlit==1.54.0
groq==1.0.0
python-dotenv==1.2.1
pymongo